+++ custom_components/supernotify/transports/alexa_devices.py
@@ -143,12 +143,13 @@ class AlexaDevicesTransport(Transport):
     async def deliver(self, envelope: Envelope, debug_trace: DebugTrace | None = None) -> bool:
         _LOGGER.debug("SUPERNOTIFY notify_alexa_devices: %s", envelope.message)
 
-        target_data: dict[str, Any] = self.action_target(envelope, envelope.target.entity_ids or None)
+        targets = envelope.target.entity_ids or []
 
-        if not self.has_action_target(target_data):
+        if not targets:
             _LOGGER.debug("SUPERNOTIFY Skipping alexa devices, no targets")
             return False
 
         action_data: dict[str, Any] = {ATTR_MESSAGE: envelope.message or ""}
+        target_data: dict[str, Any] = {ATTR_ENTITY_ID: targets}
 
         return await self.call_action(envelope, action_data=action_data, target_data=target_data)